[Chore] #3 - 검색 UI 구조 변경
SearchVC 하나 파일에 구현
